requirements.txt
scipy
bs4
tensorflow-gpu
scikit-image
tqdm
lxml
datasets==1.2.1
rouge_score
